pyproject.toml
[project]
name = "blastweb"
version = "0.1.4"
description = "A simple web UI and API for running BLAST+ against local databases."
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
  {name = "Hiroyo NISHIDE", email = "hiroyo@nibb.ac.jp"}
]
dependencies = [
  "Flask",
  "click",
  "PyYAML"
]

[project.scripts]
blastweb = "blastweb.cli:cli"

